+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [12.4.0](https://github.com/ChasarooniZ/pf2e-rpg-numbers/compare/12.3.1...12.4.0) - Start of some nice UI updates
+- `Finishing Moves`
+  - Added support for renaming items for the purposes of Finishing Moves with the menu item in the item's header
+  - More to come soon TM
+- `Version Support`
+  - Opened up allowing use of module in V11 (should be compatible as long as sequencer is compatible between the two)
+
 # [12.3.1](https://github.com/ChasarooniZ/pf2e-rpg-numbers/compare/12.3.0...12.3.1) - Fixing some Issues
 
 - `Finishing Moves`